fluxor implementation

+++ ChatApp.Server/Security/ConnectedUser.cs
@@ -8,13 +8,13 @@ public class ConnectedUser(IHttpContextAccessor httpContextAccessor) : IConnecte
 {
     private readonly IHttpContextAccessor _httpContextAccessor = httpContextAccessor;
     
-    public int? Id {
+    public long? Id {
         get
         {
             var user = _httpContextAccessor.HttpContext?.User;
             var claim = user?.Claims.FirstOrDefault(c => c.Type == JwtRegisteredClaimNames.Sid);
             if (claim is null) return null;
-            return int.Parse(claim.Value);
+            return long.Parse(claim.Value);
         }
     }
 }
